% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Utility check function for doing checks}
\usage{
check(input, func)
}
\arguments{
\item{input}{any file needed for the function you are doing checks on}

\item{func}{any function needed to run the function on the input file}
}
\value{
check_object results
}
\description{
This is an all-purpose checker to make checks based on any available check functionality we have on any datasets
}
