% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_eos_status}
\alias{check_eos_status}
\title{Check EOS status}
\usage{
check_eos_status(
  household_data,
  individual_data,
  household_data_status,
  individual_data_status
)
}
\description{
Check if household is eos then all individuals in that household needs to be eos
}
