% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthecon.R
\name{check_healthecon}
\alias{check_healthecon}
\title{Entry function for checking healtheconomics metadata}
\usage{
check_healthecon(zip_path)
}
\arguments{
\item{zip_path}{this is where your zip file is located locally}
}
\value{
returns a \code{check_results_obj}
}
\description{
this function will run all the necessary checks for health economics metadata.
}
