% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_out_status}
\alias{check_out_status}
\title{Check OUT status}
\usage{
check_out_status(
  household_data,
  individual_data,
  household_data_status,
  individual_data_status
)
}
\description{
Check if household is out then all individuals in that household needs to be out
}
