% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_zip_files_naming_validity}
\alias{check_zip_files_naming_validity}
\title{Check naming in zip file}
\usage{
check_zip_files_naming_validity(map_obj, ...)
}
\arguments{
\item{map}{object from map_metadata_staging_zip_file}
}
\value{
TRUE if ERROR is intentionally captured, FALSE if passed the test
}
\description{
Check naming in zip file
}
