% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_zip_n_files}
\alias{check_zip_n_files}
\title{Check number of contents in a zip file}
\usage{
check_zip_n_files(map_obj, ...)
}
\arguments{
\item{map}{object from map_metadata_staging_zip_file}
}
\value{
TRUE if ERROR is intentionally captured, FALSE if passed the test
}
\description{
Check number of contents in a zip file
}
