% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promote.R
\name{promote}
\alias{promote}
\title{Promote staging file to prod}
\usage{
promote(check_results_obj, store_historical = TRUE)
}
\arguments{
\item{check_results_obj}{this is a \code{check_result_object} coming from \code{check_} operation}

\item{store_historical}{parameter whether to store data as a historical file with partition}
}
\description{
Function to promote staging file to prod
}
